# etcdv2: encode Event just once before distribute to watchers