@xiang90 2016-12-13
Can you keep this change set minimal? This commit changes more than just what described in the commit message.

Also do you have any benchmark numbers to show the improvement?