@mqliang 2016-12-13
> Can you keep this change set minimal? This commit changes more than just what described in the commit message.

The reason I move `const StoreKeysPrefix ` and `const StoreKeysPrefix ` form `etcdserver/server.go` to `store/store.go` is that when `watch_hub` encode Event, it need "trimEventPrefix", I move the two const ro `store/store.go` to aviod cycle import. So, it a part of this commit.