@xiang90 2016-12-14
Can you get some benchmark number?

To be honest, we are not super interested in optimizing etcd2. Do you mind sharing your goal here?