@xiang90 2016-12-14
@mqliang Similar issue exists for etcd3 though. We would like someone to investigate into that. https://github.com/coreos/etcd/issues/4627.

1. measure overhead
2. try to fix it if the overhead is significant.