@mqliang 2016-12-14
> Can you get some benchmark number?

I benchmarked 100, 1000, 10000, 100000, 1000000 watchers, result is as follows:

```
100
BenchmarkEventMarsh-4                  	2000000000	         0.00 ns/op	       0 B/op	       0 allocs/op
BenchmarkEventMarsh-4                  	2000000000	         0.00 ns/op	       0 B/op	       0 allocs/op


1000
before:
BenchmarkEventMarsh-4                  	2000000000	         0.01 ns/op	       0 B/op	       0 allocs/op
after:
BenchmarkEventMarsh-4                  	2000000000	         0.01 ns/op	       0 B/op	       0 allocs/op

10000
before:
BenchmarkEventMarsh-4                  	1000000000	         0.11 ns/op	       0 B/op	       0 allocs/op
after:
BenchmarkEventMarsh-4                  	2000000000	         0.08 ns/op	       0 B/op	       0 allocs/op

100000
before:
BenchmarkEventMarsh-4                  	1	         		1791906046 ns/op	217701472 B/op	 1097620 allocs/op
after
BenchmarkEventMarsh-4                  	1	         		1860174846 ns/op	384462144 B/op	 1099354 allocs/op


1000000
before
BenchmarkEventMarsh-4                  	1	         		20795525659 ns/op	2177594664 B/op	10995767 allocs/op
after
BenchmarkEventMarsh-4                  	1	         		50022956594 ns/op	3842379144 B/op	11000028 allocs/op

```

Seems only make sense if we have a huge number of watcher(>1 millions)