@mqliang 2016-12-14
Post my benchmark code for reference:

``` golang
package store

import (
	"sync"
	"testing"
	"encoding/json"
)

func BenchmarkEventMarsh(b *testing.B) {
	hub := newWatchHub(1000)

	e := &Event{
		Action: Create,
		Node: &NodeExtern{
			Key:           "foo",
			ModifiedIndex: 1,
			CreatedIndex:  1,
		},
		EtcdIndex: 1,
		Refresh: false,
	}

	b.StopTimer()
	b.StartTimer()

	hub.notify(e)

	const num = 1000000

	wg := sync.WaitGroup{}
	wg.Add(num)
	for i := 0; i < num; i++ {
		go func() {
			defer wg.Done()
			w, err := hub.watch("foo", false, false, 1, 1)
			if err != nil {

				b.Fatal(err)
			}

			select {
			case e := <- w.EventChan():
				//before
				data, _ := json.Marshal(e)
				var _ = data
				//after
				// var _ = e

			}
		}()
	}
	wg.Wait()
}

```