@mqliang 2016-12-14
@xiang90 feel free to close this if my benchmark code is right.